+++ src/app/daos/[name]/mission/PageClient.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+export const runtime = "edge";
+
+export default function PageClient() {
+  // This component is just a placeholder since the actual content
+  // is rendered in the layout-client.tsx file when isMission is true
+  return null;
+}
