+++ src/app/daos/[name]/mission/page.tsx
@@ -1,9 +1,7 @@
-"use client";
-
 export const runtime = "edge";
 
-export default function MissionPage() {
-  // This component is just a placeholder since the actual content
-  // is rendered in the layout-client.tsx file when isMission is true
-  return null;
+import PageClient from "./PageClient";
+
+export default async function Page() {
+  return <PageClient />;
 }
