+++ src/app/page.tsx
@@ -1,4 +1,5 @@
 import Home from "@/components/home/Home";
-export default function page() {
+
+export default async function Page() {
   return <Home />;
 }
